double/float semantics changes for Win32
r93

+++ rtlib/libfb_str_convfrom.c
@@ -103,7 +103,7 @@ FBCALL double fb_hStr2Double( char *src, int len )
 			return (double)fb_hStrToInt( &p[2], len-2, radix );
 	}
 	else
-		return (double)atof( p );
+		return atof( p );
 
 }
